\name{monomvn-package}
\alias{monomvn-package}
\docType{package}
\title{Estimation for Multivariate Normal Data with Monotone Missingness}
\description{
Estimation of multivariate normal data of arbitrary dimension
where the pattern of missing data is monotone. Through the use of
partial least squares and principal component regressions, 
where standard regressions fail,
the package can handle an (almost) arbitrary amount of missing
data.  The current version supports maximum likelihood inference.
Future versions will provide a means of sampling from a Bayesian
posterior.
}
\details{
For a fuller overview including a complete list of functions, demos and
vignettes, please use \code{help(package="tgp")}.
}
}
\author{
Robert B. Gramacy \email{bobby@statslab.cam.ac.uk}

Maintainer: 
Robert B. Gramacy \email{bobby@statslab.cam.ac.uk}
}
\references{
 \emph{Statistical Analysis with Missing Data}, Second Edition.
   Roderick J.A. Little and Donald B. Rubin.  Wilely. (2002)

}
\keyword{ package }
\seealso{ \code{\link{monomvn}}, \pkg{norm}, \pkg{mvnmle} }
}

