\name{monomvn-internal}
%% Part of the sensible export list but (currently?) documented as
%% internal (or waiting for documentation to be written):
%% <LIST>
\alias{addy}
\alias{regress}

\title{Internal Monomvn Functions}
\description{
  Internal monomvn functions
}
\details{
  These are not to be called by the user (or in some cases are just
  waiting for proper documentation to be written :)).
}
%\references{
%\url{http://www.ams.ucsc.edu/~rbgramacy/tgp.html}
%}
\keyword{ internal }
