\name{monomvn-internal}
%% Part of the sensible export list but (currently?) documented as
%% internal (or waiting for documentation to be written):
%% <LIST>
\alias{my.Igamma.inv}
\alias{bmonomvn.read.traces}
\alias{addy}
\alias{regress.ls}
\alias{regress.lars}
\alias{regress.pls}
\alias{regress.ridge}
\alias{opt.ridge}

\title{Internal Monomvn Functions}
\description{
  Internal monomvn functions
}
\details{
  These are not to be called by the user (or in some cases are just
  waiting for proper documentation to be written :)).
}
%\references{
%\url{http://www.statslab.cam.ac.uk/~bobby/monomvn.html}
%}
\keyword{ internal }
