\name{monomvn-internal}
%% Part of the sensible export list but (currently?) documented as
%% internal (or waiting for documentation to be written):
%% <LIST>
\alias{Igamma.inv}
\alias{get.lambda}
\alias{get.regress}
\alias{get.regress.C}
\alias{adjust.elist.C}
\alias{mvnpdf.C}
\alias{bmonomvn.read.traces}
\alias{read.DA.trace}
\alias{read.muS.trace}
\alias{addy}
\alias{regress.ls}
\alias{regress.lars}
\alias{regress.pls}
\alias{regress.ridge}
\alias{regress.fact}
\alias{opt.ridge}
\alias{blasso.cleanup}
\alias{bmonomvn.cleanup}
\alias{check.start}
\alias{check.QP}
\alias{postprocess.QP}
\alias{table2blasso}
\alias{da.perm}
\alias{add.pe.QP}
\alias{bhs}

\title{Internal Monomvn Functions}
\description{
  Internal monomvn functions
}
\details{
  These are not to be called by the user (or in some cases are just
  waiting for proper documentation to be written :))
}
%\references{
%\url{http://www.statslab.cam.ac.uk/~bobby/monomvn.html}
%}
\keyword{ internal }
